% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.R
\name{get.label.name.list}
\alias{get.label.name.list}
\title{Get Stata Variable labels}
\usage{
get.label.name.list(dat, lang = NA)
}
\arguments{
\item{dat}{data.frame. Imported data.frame}

\item{lang}{string. Label language. Default language defined by get.lang() is used if NA}

\item{var.name}{character. Variable name}
}
\value{
Returns an named vector of variable labels
}
\description{
Get Stata Variable labels
}

