% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.label.name}
\alias{get.label.name}
\title{Get name of Stata label set for a variable}
\usage{
get.label.name(x, var.name)
}
\arguments{
\item{x}{data.frame. Imported data.frame}

\item{var.name}{character. Variable name}
}
\description{
Get name of Stata label set for a variable
}

