% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.R
\name{get.label}
\alias{get.label}
\title{Get Stata label table for a label set}
\usage{
get.label(dat, label.name)
}
\arguments{
\item{dat}{data.frame. Data.frame created by read.stata13.}

\item{label.name}{character. Name of the Stata label set}
}
\value{
Returns a named vector of code numbers
}
\description{
Get Stata label table for a label set
}
\details{
This function returns the table of factor levels which represent a Stata label set.
The name of a label set for a variable can be obtained by \code{\link{get.label.name}}.
}
\examples{
dat <- read.dta13("http://www.stata-press.com/data/r13/auto.dta")
labname <- get.label.name(dat,"foreign")
get.label(dat, labname)
}

