% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.R
\name{get.label.table}
\alias{get.label.table}
\title{Get Stata label set for a variable}
\usage{
get.label.table(dat, label.name)
}
\arguments{
\item{dat}{data.frame. Imported data.frame}

\item{label.name}{character. Name of the Stata label set}
}
\value{
Returns a named vector of code numbers
}
\description{
Get Stata label set for a variable
}

