% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get.label.tables}
\alias{get.label.tables}
\title{Get all Stata Label Sets for a Data.frame}
\usage{
get.label.tables(dat)
}
\arguments{
\item{dat}{\emph{data.frame.} Data.frame created by \code{read.dta13}.}
}
\value{
Returns a named list of label tables
}
\description{
Retrieve the value labels for all variables.
}
\details{
This function returns the factor levels which represent
 a Stata label set for all variables.
}
\examples{
dat <- read.dta13(system.file("extdata/statacar.dta", package="readstata13"))
get.label.tables(dat)
}
\author{
Jan Marvin Garbuszus \email{jan.garbuszus@ruhr-uni-bochum.de}

Sebastian Jeworutzki \email{sebastian.jeworutzki@ruhr-uni-bochum.de}
}
