% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.R
\name{get.lang.list}
\alias{get.lang.list}
\title{List label languages}
\usage{
get.lang.list(dat)
}
\arguments{
\item{dat}{data.frame. Imported data.frame}
}
\value{
Returns a vector of language labels
}
\description{
List label languages
}

