% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.R
\name{get.origin.codes}
\alias{get.origin.codes}
\title{Get Origin Code Numbers for Factors}
\usage{
get.origin.codes(x, label.table)
}
\arguments{
\item{x}{\emph{factor.} Factor to obtain code for}

\item{label.table}{\emph{table.} Table with factor levels obtained by \code{\link{get.label}}.}
}
\value{
Returns an integer with original codes
}
\description{
Recreates the code numbers of a factor as stored in the Stata dataset.
}
\details{
While converting numeric variables to factors, the original code numbers is lost.  This function reconstructs the codes from the attribute "label.table".
}
\examples{
dat <- read.dta13("http://www.stata-press.com/data/r13/auto.dta")
labname <- get.label.name(dat,"foreign")
labtab <- get.label(dat, labname)

# comparsion
get.origin.codes(dat$foreign, labtab)
as.integer(dat$foreign)
}
\author{
Jan Marvin Garbuszus \email{jan.garbuszus@ruhr-uni-bochum.de}

Sebastian Jeworutzki \email{sebastian.jeworutzki@ruhr-uni-bochum.de}
}

