% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.R
\name{get.varlabel}
\alias{get.varlabel}
\title{Get Stata Variable labels}
\usage{
get.varlabel(dat, var.name = NULL, lang = NA)
}
\arguments{
\item{dat}{data.frame. Data.frame created by read.stata13.}

\item{var.name}{character vector. Variable names. If NULL label for all variables.}

\item{lang}{string. Label language. Default language defined by \code{\link{get.lang}} is used if NA}
}
\value{
Returns an named vector of variable labels
}
\description{
Get Stata Variable labels
}

