% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{maxchar}
\alias{maxchar}
\title{Check max char length of data.frame vectors}
\usage{
maxchar(x)
}
\arguments{
\item{x}{vector of data frame}
}
\description{
Stata requires us to provide the maximum size of a charactervector as every
row is stored in a bit region of this size.
}
\details{
Ex: If the max chars size is four, _ is no character in this vector:
1. row: four
3. row: one_
4. row: ____

If a character vector contains only missings or is empty, we will assign it a
value of one, since Stata otherwise cannot handle what we write.
}
