% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.dta13}
\alias{read.dta13}
\title{Read Stata 13 Binary Files}
\usage{
read.dta13(path, convert.factors = TRUE)
}
\arguments{
\item{path}{The dta file you want to import.}

\item{convert.factors}{logical Use Stata value labels to create factors?
(Version 6.0 or later).}
}
\value{
The function returns a data.frame with attributs. The attributes include
\describe{
  \item{datalabel}{Dataset label}
  \item{time.stamp}{Timestamp of file creation}
  \item{formats}{Stata display formats. May be used with \code{\link{sprintf}}}
  \item{type}{Stata data type (see Stata Corp 2014)}
  \item{val.labels}{For each variable the name of the associated value labels in "label"}
  \item{var.labels}{Variable labels}
  \item{version}{dta file format version}
  \item{label}{List of value labels: ["labelname"]["code"] is a numeric vector with codes and ["labelname"]["label"] ia a character vector with label text. }
  \item{strl}{List of character vectors for the new strl string variable type. The first element is the identifier, and the second element the string.}
}
}
\description{
\code{read.dta13} reads a Stata 13 dta file bytewise and imports the data
into a data.frame.
}
\note{
If you catch a bug, please do not sue us, we do not have any money.
}
\author{
Jan Marvin Garbuszus \email{jan.garbuszus@rub.de}

Sebastian Jeworutzki \email{sebastian.jeworutzki@rub.de}
}
\references{
Stata Corp (2014): Description of .dta file format \url{http://www.stata.com/help.cgi?dta}
}
\seealso{
\code{\link{read.dta}} and \code{memisc} for dta files from Stata
Versions < 13
}

