% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.R
\name{read.encoding}
\alias{read.encoding}
\title{Wrapper around iconv calls for code readability}
\usage{
read.encoding(x, encoding)
}
\arguments{
\item{x}{element to be converted}

\item{encoding}{encoding to be used.}
}
\description{
Wrapper around iconv calls for code readability
}

