% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{readstata13}
\alias{readstata13}
\alias{readstata13-package}
\title{Import Stata 13 Datafiles}
\description{
Function to read the Stata 13 file format into a data.frame.
}
\note{
If you catch a bug, please do not sue us, we do not have any money.
}
\author{
Marvin Garbuszus \email{jan.garbuszus@ruhr-uni-bochum.de},
Sebastian Jeworutzki \email{Sebastian.Jeworutzki@ruhr-uni-bochum.de}
}
\seealso{
\code{\link{read.dta}} and \code{memisc} for dta files from Stata
Versions < 13
}

