% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{saveToExport}
\alias{saveToExport}
\title{Check if numeric vector can be expressed as interger vector
Compression can reduce numeric vectors as integers if the vector does only
contain integer type data.}
\usage{
saveToExport(x)
}
\arguments{
\item{x}{vector of data frame}
}
\description{
Check if numeric vector can be expressed as interger vector
Compression can reduce numeric vectors as integers if the vector does only
contain integer type data.
}

