% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.R
\name{set.label}
\alias{set.label}
\title{Assign Stata labels to a variable}
\usage{
set.label(x, var.name)
}
\arguments{
\item{x}{data.frame. Imported data.frame}

\item{var.name}{character. Name of the variable in the data.frame}
}
\value{
Returns a labeled factor
}
\description{
Assign Stata labels to a variable
}
\examples{
dat <- read.dta13("http://www.stata-press.com/data/r13/auto.dta",
                   convert.factors = FALSE)

# compare vectors
set.label(dat, "foreign")
dat$foreign
}

