% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.R
\name{set.lang}
\alias{set.lang}
\title{Assign Stata Language Labels}
\usage{
set.lang(dat, lang = NA, generate.factors = FALSE)
}
\arguments{
\item{dat}{\emph{data.frame.} Data.frame created by read.dta13.}

\item{lang}{\emph{character.} Label language. Default language defined by \code{\link{get.lang}} is used if NA}

\item{generate.factors}{\emph{logical.} If TRUE generates factors.}
}
\value{
Returns a data.frame with value labels in language "lang".
}
\description{
Changes default label language for a dataset.
}
\author{
Jan Marvin Garbuszus \email{jan.garbuszus@ruhr-uni-bochum.de}

Sebastian Jeworutzki \email{sebastian.jeworutzki@ruhr-uni-bochum.de}
}

