% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stata}
\alias{stata}
\title{Reads the binary Stata file}
\usage{
stata(filePath, missing)
}
\arguments{
\item{filePath}{The full systempath to the dta file you want to import.}

\item{missing.type}{logical if missings should be converted outside of Rcpp.}
}
\description{
Reads the binary Stata file
}
