% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stata_pre13_save}
\alias{stata_pre13_save}
\title{Writes the binary Stata file}
\usage{
stata_pre13_save(filePath, dat)
}
\arguments{
\item{filePath}{The full systempath to the dta file you want to export.}

\item{dat}{an R-Object of class data.frame.}
}
\description{
Writes the binary Stata file
}

