% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stataWrite}
\alias{stataWrite}
\title{Reads the binary Stata file}
\usage{
stataWrite(filePath, dat)
}
\arguments{
\item{filePath}{The full systempath to the dta file you want to import.}

\item{dat}{an R-Object of class data.frame.}
}
\description{
Reads the binary Stata file
}

