% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stataWriteOld}
\alias{stataWriteOld}
\title{Writes the binary Stata file}
\usage{
stataWriteOld(filePath, dat)
}
\arguments{
\item{filePath}{The full systempath to the dta file you want to export.}

\item{dat}{an R-Object of class data.frame.}
}
\description{
Writes the binary Stata file
}

