% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.label.table}
\alias{get.label.table}
\title{Get Stata label set for a variable}
\usage{
get.label.table(x, label.name)
}
\arguments{
\item{x}{data.frame. Imported data.frame}

\item{label.name}{character. Name of the Stata label set}
}
\description{
Get Stata label set for a variable
}

