% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.R
\name{get.lang}
\alias{get.lang}
\title{Show default label language}
\usage{
get.lang(dat)
}
\arguments{
\item{dat}{data.frame. Imported data.frame}
}
\value{
If a language label was set, it will be returned otherwise NA.
}
\description{
Show default label language
}

