% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get.lang}
\alias{get.lang}
\title{Show Default Label Language}
\usage{
get.lang(dat, print = T)
}
\arguments{
\item{dat}{\emph{data.frame.} Data.frame created by \code{read.dta13}.}

\item{print}{\emph{logical.} If \code{TRUE}, print available languages and
default language.}
}
\value{
Returns a list with two components:
\describe{
\item{languages:}{Vector of label languages used in the dataset}
\item{default:}{Name of the actual default label language, otherwise NA}
}
}
\description{
Displays informations about the defined label languages.
}
\details{
Stata allows to define multiple label sets in different languages.
 This functions reports the available languages and the selected default
 language.
}
\author{
Jan Marvin Garbuszus \email{jan.garbuszus@ruhr-uni-bochum.de}

Sebastian Jeworutzki \email{sebastian.jeworutzki@ruhr-uni-bochum.de}
}
