% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.origin.codes}
\alias{get.origin.codes}
\title{Get original code number}
\usage{
get.origin.codes(x, label.table)
}
\arguments{
\item{x}{factor. Factor to obtain code for}

\item{label.table}{table. Table with factor levels - obtained by get.labeltable()}
}
\description{
Get original code number
}
\examples{
dat <- read.dta13("http://www.stata-press.com/data/r13/auto.dta")
labname <- get.label.name(dat,"foreign")
labtab <- get.label.table(dat, labname)

# comparsion
get.origin.codes(dat$foreign, labtab)
as.integer(dat$foreign)
}

