% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.R
\name{get.varlabel}
\alias{get.varlabel}
\title{Get Stata Variable Labels}
\usage{
get.varlabel(dat, var.name = NULL, lang = NA)
}
\arguments{
\item{dat}{\emph{data.frame.} Data.frame created by \code{read.dta13}.}

\item{var.name}{\emph{character vector.} Variable names. If NULL, get label for all variables.}

\item{lang}{\emph{character.} Label language. Default language defined by \code{\link{get.lang}} is used if NA}
}
\value{
Returns an named vector of variable labels
}
\description{
Retrieve variable labels from dataset attributes.
}
\author{
Jan Marvin Garbuszus \email{jan.garbuszus@ruhr-uni-bochum.de}

Sebastian Jeworutzki \email{sebastian.jeworutzki@ruhr-uni-bochum.de}
}

