% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.dta13}
\alias{read.dta13}
\title{Read Stata 13 Binary Files}
\usage{
read.dta13(path, convert.factors = TRUE, fileEncoding = NULL)
}
\arguments{
\item{path}{string path to the dta file you want to import}

\item{convert.factors}{logical create factors from Stata value labels}

\item{fileEncoding}{string If not null, strings will be converted from fileEncoding to system encoding}
}
\value{
The function returns a data.frame with attributs. The attributes include
\describe{
  \item{datalabel}{Dataset label}
  \item{time.stamp}{Timestamp of file creation}
  \item{formats}{Stata display formats. May be used with \code{\link{sprintf}}}
  \item{type}{Stata data type (see Stata Corp 2014)}
  \item{val.labels}{For each variable the name of the associated value labels in "label"}
  \item{var.labels}{Variable labels}
  \item{version}{dta file format version}
  \item{lable.table}{List of value labels.}
  \item{strl}{List of character vectors for the new strl string variable type. The first element is the identifier and the second element the string.}
  \item{expansion.field}{list providing variable name, characteristic name and the contents of stata characteristic field.}
}
}
\description{
\code{read.dta13} reads a Stata 13 dta file bytewise and imports the data
into a data.frame.
}
\note{
If you catch a bug, please do not sue us, we do not have any money.
}
\author{
Jan Marvin Garbuszus \email{jan.garbuszus@rub.de}

Sebastian Jeworutzki \email{sebastian.jeworutzki@rub.de}
}
\references{
Stata Corp (2014): Description of .dta file format \url{http://www.stata.com/help.cgi?dta}
}
\seealso{
\code{\link{read.dta}} and \code{memisc} for dta files from Stata
Versions < 13
}

