% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readstata13.R
\docType{package}
\name{readstata13}
\alias{readstata13}
\alias{readstata13-package}
\title{Import Stata Data Files}
\description{
Function to read the Stata file format into a data.frame.
}
\note{
If you catch a bug, please do not sue us, we do not have any money.
}
\seealso{
\code{\link[foreign]{read.dta}} and \code{memisc} for dta files from Stata
Versions < 13
}
\author{
Marvin Garbuszus \email{jan.garbuszus@ruhr-uni-bochum.de}

Sebastian Jeworutzki \email{sebastian.jeworutzki@ruhr-uni-bochum.de}
}
