% Generated by roxygen2 (4.0.2): do not edit by hand
\name{save.dta13}
\alias{save.dta13}
\title{Write Stata 13 Binary Files}
\usage{
save.dta13(data, file = "path", data.label = NULL, time.stamp = TRUE,
  convert.factors = FALSE, convert.dates = TRUE)
}
\arguments{
\item{data}{data.frame. A data.frame Object.}

\item{data.label}{string. Name of the dta-file.}

\item{time.stamp}{logical. If TRUE add a time.stamp to the dta-file.}

\item{path}{string. Path to the dta file you want to export.}
}
\value{
The function writes a dta-file to disk. The following features of the dta file format are supported:
\describe{
  \item{datalabel:}{Dataset label}
  \item{time.stamp:}{Timestamp of file creation}
  \item{formats:}{Stata display formats. May be used with \code{\link{sprintf}}}
  \item{type:}{Stata data type (see Stata Corp 2014)}
  \item{var.labels:}{Variable labels}
  \item{version:}{dta file format version}
  \item{strl:}{List of character vectors for the new strL string variable type. The first element is the identifier and the second element the string.}
}
}
\description{
\code{save.dta13} writes a Stata 13 dta file bytewise and saves the data
into a dta-file.
}
\author{
Jan Marvin Garbuszus \email{jan.garbuszus@rub.de}

Sebastian Jeworutzki \email{sebastian.jeworutzki@rub.de}
}
\references{
Stata Corp (2014): Description of .dta file format \url{http://www.stata.com/help.cgi?dta}
}
\seealso{
\code{\link{write.dta}} and \code{memisc} for dta files from Stata
versions < 13.
}

