% Generated by roxygen2 (4.0.2): do not edit by hand
\name{save.dta13}
\alias{save.dta13}
\title{Write Stata 13 Binary Files}
\usage{
save.dta13(data, file = "path", data.label = NULL, time.stamp = TRUE,
  convert.factors = FALSE, convert.dates = TRUE, tz = "GMT",
  add.rownames = FALSE)
}
\arguments{
\item{data}{data.frame. A data.frame Object.}

\item{data.label}{string. Name of the dta-file.}

\item{time.stamp}{logical. If TRUE add a time.stamp to the dta-file.}

\item{convert.factors}{logical. If TRUE factors will be converted to Stata variables with labels. Stata expects strings to be encoded as Windows-1252, so all levels will be recoded.  Character which can not be mapped in Windows-1252 will be saved as hexcode.}

\item{convert.dates}{logical. If TRUE dates will be converted to Stata date time format. Code from foreign::write.dta()}

\item{tz}{string. The name of the timezone convert.dates will use.}

\item{add.rownames}{logical. If TRUE a new variable rownames will be added to the dta-file.}

\item{path}{string. Path to the dta file you want to export.}
}
\value{
The function writes a dta-file to disk. The following features of the dta file format are supported:
\describe{
  \item{datalabel:}{Dataset label}
  \item{time.stamp:}{Timestamp of file creation}
  \item{formats:}{Stata display formats. May be used with \code{\link{sprintf}}}
  \item{type:}{Stata data type (see Stata Corp 2014)}
  \item{var.labels:}{Variable labels}
  \item{version:}{dta file format version}
  \item{strl:}{List of character vectors for the new strL string variable type. The first element is the identifier and the second element the string.}
}
}
\description{
\code{save.dta13} writes a Stata 13 dta file bytewise and saves the data
into a dta-file.
}
\author{
Jan Marvin Garbuszus \email{jan.garbuszus@rub.de}

Sebastian Jeworutzki \email{sebastian.jeworutzki@rub.de}
}
\references{
Stata Corp (2014): Description of .dta file format \url{http://www.stata.com/help.cgi?dta}
}
\seealso{
\code{\link{write.dta}} and \code{memisc} for dta files from Stata
versions < 13.
}

