% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.R
\name{set.label}
\alias{set.label}
\title{Assign Stata labels to a variable}
\usage{
set.label(dat, var.name, lang = NA)
}
\arguments{
\item{dat}{data.frame. Data.frame created by read.stata13.}

\item{var.name}{character. Name of the variable in the data.frame}

\item{lang}{string. Label language. Default language defined by \code{\link{get.lang}} is used if NA}
}
\value{
Returns a labeled factor
}
\description{
Assign Stata labels to a variable
}
\examples{
dat <- read.dta13("http://www.stata-press.com/data/r13/autofull.dta",
                   convert.factors = FALSE)

# compare vectors
set.label(dat, "foreign")
dat$foreign
}

