% Generated by roxygen2 (4.0.2): do not edit by hand
\name{set.label}
\alias{set.label}
\title{Assign Stata labels to a variable}
\usage{
set.label(x, var.name)
}
\arguments{
\item{x}{data.frame. Imported data.frame}

\item{var.name}{character. Name of the variable in the data.frame}
}
\description{
Assign Stata labels to a variable
}
\examples{
dat <- read.dta13("http://www.stata-press.com/data/r13/auto.dta",
                   convert.factors=F)

# compare vectors
set.label(dat, "foreign")
dat$foreign
}

