% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stata}
\alias{stata}
\title{Reads the binary Stata file}
\usage{
stata(filePath, missing)
}
\arguments{
\item{filePath}{The full systempath to the dta file you want to import.}

\item{missing}{logical if missings should be converted outside of Rcpp.}
}
\description{
Reads the binary Stata file
}

