% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stataWrite}
\alias{stataWrite}
\title{Writes the binary Stata file}
\usage{
stataWrite(filePath, dat)
}
\arguments{
\item{filePath}{The full systempath to the dta file you want to export.}

\item{dat}{an R-Object of class data.frame.}
}
\description{
Writes the binary Stata file
}

