% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{varlabel}
\alias{varlabel}
\alias{varlabel<-}
\title{Get and assign Stata Variable Labels}
\usage{
varlabel(dat, var.name = NULL, lang = NA)

varlabel(dat) <- value
}
\arguments{
\item{dat}{\emph{data.frame.} Data.frame created by \code{read.dta13}.}

\item{var.name}{\emph{character vector.} Variable names. If NULL, get label for all variables.}

\item{lang}{\emph{character.} Label language. Default language defined by \code{\link{get.lang}} is used if NA}

\item{value}{\emph{character vector.} Vector of variable names.}
}
\value{
Returns an named vector of variable labels
}
\description{
Retrieve or set variable labels for a dataset.
}
\author{
Jan Marvin Garbuszus \email{jan.garbuszus@ruhr-uni-bochum.de}

Sebastian Jeworutzki \email{sebastian.jeworutzki@ruhr-uni-bochum.de}
}

